core/UsraWise/requirements.txt
langchain
google-generativeai
langchain-google-genai
streamlit
python-dotenv
langchain-community
